requirements.txt
streamlit
langchain
langchain-community
pypdf
sentence-transformers
numpy
arxiv
faiss-cpu
lxml
langchain-openai==1.1.7
openai==2.15.0
tiktoken==0.12.0
distro==1.9.0
jiter==0.12.0
sniffio==1.3.1
langchain-text-splitters==1.1.0
langchain-core==1.2.7
langchain-classic==1.0.1
